Generated test methods for Collection fields
Getter methods for Collection fields have the side-effect that they
initialize a previously unititialized (ie. null) field with an empty
Collection. This leads to unnecessary memory consumption.
The workaround is to test if the Collection field is empty or not
before invoking the getter method. For example:

Model model = ...;
if(model.hasExtensions()){
  List<Extensions> extensions = model.getExtensions();
  for(Extension extension : extensions){
    // Do something useful with the extension element
  }
}